Separate all state machine stuff from core0.
Only needed when server calls for an update to the pedsup.
Now, core0/cdc_task sends an enum CMD to core1 via the
native pico queue. Core 1, in response, begins its loop
checking for anything on the queue.

If this pattern works, GPIO pin manip is next, and the
sm_array can be removed from the shared state.