Catch disconnected picos (#9)
* Catch disconnected picos

Whether a pico is connected or not is extern registered in lv-server init. Pico
status is then checked in the handler -- commands to disconnected picos no
longer added to the queue (which was bricking the server). ERROR sent to the
client which catches it with a TypeError. One possible improvement: recheck 
somehow if the pico becomes connected after init.

* Explicitly specify bool output formats to trip pico disconnect error.

* Remove mac cache file.